bookstore/requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.5.3
dj-database-url==2.2.0
Django==5.0.4
django-cors-headers==4.3.1
django-decouple==2.1
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==22.0.0
iniconfig==2.0.0
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
